==> models/6_Hyperparameter_Tuning.ipynb <==
<div align="center">
    <h1><b><u>TRAFFIC OPTIMIZATION</u></b></h1>
</div>


**GROUP NUMBER: 7**
     
**GROUP MEMBERS**
   
- **Adebola**
- **Rahim**
- **Sayeed**
- **Yinka**
- **Minto**
   

<div align="center">
    <h3><b><u>6. Hyper-Parameter Tuning</u></b></h3>
</div>


# import libraries 
import pandas as pd
from sklearn.ensemble import RandomForestRegressor
from sklearn.model_selection import GridSearchCV
from sklearn.model_selection import train_test_split

# load dataset
df = pd.read_csv('data/encoded_data.csv')
df.head()

# Separate data into features and target
X = df[['INVAGE', 'ACCLASS_Non-Fatal Injury', 'ACCLASS_Property Damage O', 'DIVISION', 'IMPACTYPE', 'INVTYPE',
    'LIGHT', 'INITDIR', 'TRAFFCTL', 'ROAD_CLASS', 'DRIVACT']]
y = df['INJURY']

# Split data into training and test 
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2)

# Define the base model
rf = RandomForestRegressor(random_state=42)

# Define the parameter grid
param_grid = {
    'n_estimators': [100, 200],
    'max_depth': [None, 10, 20],
    'min_samples_split': [2, 5],
    'min_samples_leaf': [1, 2],
    'max_features': ['sqrt', 'log2'],
    'bootstrap': [True, False]
}

# Grid search
grid_search = GridSearchCV(estimator=rf,
                           param_grid=param_grid,
                           cv=3,
                           scoring='neg_mean_absolute_error',
                           n_jobs=-1,
                           verbose=2)

grid_search.fit(X_train, y_train)

# Best model
best_rf = grid_search.best_estimator_
print("Best Parameters:", grid_search.best_params_)
