==> models/7_Random_Forest.ipynb <==
<div align="center">
    <h1><b><u>TRAFFIC OPTIMIZATION</u></b></h1>
</div>


**GROUP NUMBER: 7**
     
**GROUP MEMBERS**
   
- **Adebola**
- **Rahim**
- **Sayeed**
- **Yinka**
- **Minto**
   

<div align="center">
    <h3><b><u>7. Random Forest</u></b></h3>
</div>


### 1. Data Load

# Import Libraries 
import pandas as pd
import numpy as np
from sklearn.pipeline import Pipeline
from sklearn.compose import ColumnTransformer
from sklearn.preprocessing import FunctionTransformer, OneHotEncoder
from sklearn.ensemble import RandomForestClassifier
from sklearn.impute import SimpleImputer
from sklearn.model_selection import train_test_split
import joblib


# Load your dataset
data = pd.read_csv('data/cleaned_data.csv')

data.head()

### 2. Data Pre-processing

# Encode target
injury_mapping = {"Minimal": 0, "Minor": 1, "Major": 2, "Fatal": 3}
data["INJURY"] = data["INJURY"].map(injury_mapping)

# Feature and target definition
features = [
    "DRIVCOND", "INVTYPE", "IMPACTYPE", "ACCLASS"
]
X = data[features]
y = data["INJURY"]

# Define preprocessing
preprocessor = ColumnTransformer(transformers=[
    
    ("cat", OneHotEncoder(handle_unknown="ignore"), [
        "DRIVCOND", "INVTYPE", "IMPACTYPE", "ACCLASS"
    ])
])

# Create and train pipeline
pipeline = Pipeline(steps=[
    ("preprocessor", preprocessor),
    ("classifier", RandomForestClassifier(n_estimators=50,
    max_depth=10,
    min_samples_split=5, class_weight='balanced'))
])

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)
pipeline.fit(X_train, y_train)



from sklearn.metrics import accuracy_score, classification_report

y_pred = pipeline.predict(X_test)
print("Accuracy:", accuracy_score(y_test, y_pred))
print(classification_report(y_test, y_pred))


# Save the pipeline
#joblib.dump(pipeline, "rf_pipeline_raw_input.pkl")

import pickle

# Save the trained pipeline
with open("rf_pipeline_raw_input.pkl", "wb") as f:
    pickle.dump(pipeline, f)
